rf: Move streamFromUint8Array helper from tests, use throughout
